Added auto-grading for group feedback submissions

+++ init.sql
@@ -79,8 +79,8 @@ CREATE TABLE IF NOT EXISTS submission (
     fileType VARCHAR(100),
     submissionDate DATETIME,
     autoGrade INT DEFAULT 0,
-    grade INT,
-    groupID INT,
+    grade DECIMAL(5, 2),
+    groupID DECIMAL(5, 2),
     FOREIGN KEY (assignmentID) REFERENCES assignment(assignmentID) ON DELETE CASCADE,
     FOREIGN KEY (studentID) REFERENCES student(studentID) ON DELETE SET NULL
 );
